requirements-dev.txt
appdirs==1.4.4
asgiref==3.7.2
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0
distlib==0.3.1
distro==1.5.0
Django==4.2.7
djangorestframework==3.14.0
html5lib==1.1
idna==3.2
lockfile==0.12.2
msgpack==1.0.2
ordered-set==4.0.2
packaging==20.9
pep517==0.10.0
Pillow==10.1.0
progress==1.5
psycopg2-binary==2.9.9
pyparsing==2.4.7
pytz==2023.3.post1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.4
toml==0.10.2
typing-extensions==4.8.0
urllib3==1.26.5
webencodings==0.5.1
appdirs==1.4.4
asgiref==3.7.2
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0
distlib==0.3.1
distro==1.5.0
Django==4.2.7
djangorestframework==3.14.0
html5lib==1.1
idna==3.2
lockfile==0.12.2
msgpack==1.0.2
ordered-set==4.0.2
packaging==20.9
pep517==0.10.0
Pillow==10.1.0
progress==1.5
psycopg2-binary==2.9.9
pyparsing==2.4.7
pytz==2023.3.post1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.4
toml==0.10.2
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.5
webencodings==0.5.1
appdirs==1.4.4
asgiref==3.7.2
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0
distlib==0.3.1
distro==1.5.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
html5lib==1.1
idna==3.2
lockfile==0.12.2
msgpack==1.0.2
ordered-set==4.0.2
packaging==20.9
pep517==0.10.0
Pillow==10.1.0
progress==1.5
psycopg2-binary==2.9.9
pyparsing==2.4.7
pytz==2023.3.post1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.4
toml==0.10.2
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.5
webencodings==0.5.1
